scripts/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
docker==7.1.0
idna==3.8
mysql-connector-python==9.0.0
pywin32==306
requests==2.32.3
urllib3==2.2.2
